move sparkParameters from spark to api
