+++ dolphinscheduler-task-plugin/dolphinscheduler-task-spark/src/test/java/org/apache/dolphinscheduler/plugin/task/spark/SparkParametersTest.java
@@ -22,6 +22,7 @@ import org.apache.dolphinscheduler.plugin.task.api.model.ResourceInfo;
 import java.util.LinkedList;
 import java.util.List;
 
+import org.apache.dolphinscheduler.plugin.task.api.parameters.SparkParameters;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
